#278 Replace `create_account` by `transfer`, `allocate` and `assign`

+++ evm_loader/program/src/account_data.rs
@@ -9,7 +9,7 @@ use solana_program::{
 /// Ethereum account version
 pub const ACCOUNT_SEED_VERSION: u8 = 1_u8;
 /// Ethereum account allocated data size
-pub const ACCOUNT_MAX_SIZE: u64 = 256;
+pub const ACCOUNT_MAX_SIZE: usize = 256;
 
 /// Ethereum account data
 #[derive(Debug,Clone)]
